refactor: consolidate useAuth hook
