+++ apps/web/core/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from "./hooks";
+export * from "./useAuth";
 export * from "./useChain";
 export * from "./usePoll";
 export * from "./useVote";
